Moves some logic from FilesRouter to FilesController for content-type and filename
