+++ src/Controllers/FilesController.js
@@ -3,6 +3,8 @@ import { Parse } from 'parse/node';
 import { randomHexString } from '../cryptoUtils';
 import AdaptableController from './AdaptableController';
 import { FilesAdapter } from '../Adapters/Files/FilesAdapter';
+import path  from 'path';
+import mime from 'mime';
 
 export class FilesController extends AdaptableController {
 
@@ -11,7 +13,19 @@ export class FilesController extends AdaptableController {
   }
 
   createFile(config, filename, data, contentType) {
+    
+    let extname = path.extname(filename);
+    
+    const hasExtension = extname.length > 0;
+    
+    if (!hasExtension && contentType && mime.extension(contentType)) {
+      filename = filename + '.' + mime.extension(contentType);
+    } else if (hasExtension && !contentType) {
+      contentType = mime.lookup(filename);
+    }
+
     filename = randomHexString(32) + '_' + filename;
+
     var location = this.adapter.getFileLocation(config, filename);
     return this.adapter.createFile(config, filename, data, contentType).then(() => {
       return Promise.resolve({
